Improve symlink script

+++ scripts/symlinks.sh
@@ -1,53 +1,55 @@
-symlink_dotfile() {
-  ln -sfv $(grealpath $1) "$HOME/.$1"
-}
-
-symlink() {
-  ln -sfv $(grealpath $1) "$HOME/$1"
-}
+ symlink() {
+   filePath=$(grealpath $1)
+   [ ! -e "$HOME/$2$1" ] && ln -sfv $filePath "$HOME/$2$1"
+ }
 
 echo "Creating symlinks ..."
 
 pushd ./dots
-symlink redis.conf
-symlink_dotfile ackrc
-symlink_dotfile agignore
-symlink_dotfile aliases
-symlink_dotfile alacritty.yml
-symlink_dotfile chunkwmrc
-symlink_dotfile ctags
-symlink_dotfile curlrc
-symlink_dotfile editorconfig
-symlink_dotfile functions
-symlink_dotfile gitconfig
-symlink_dotfile gitignore
-symlink_dotfile gitmessage
-symlink_dotfile hushlogin
-symlink_dotfile khdrc
-symlink_dotfile ripgreprc
-symlink_dotfile rgignore
-symlink_dotfile tern-project
-symlink_dotfile tmux.conf
-symlink_dotfile tmuxline_snapshot
-symlink_dotfile vimrc
-symlink_dotfile wgetrc
-symlink_dotfile zshenv
-symlink_dotfile zshrc
-ln -sfv $(grealpath tern-project) "$HOME/.tern-config"
+
+dots=(
+  ackrc
+  agignore
+  aliases
+  alacritty.yml
+  chunkwmrc
+  ctags
+  curlrc
+  editorconfig
+  functions
+  gitconfig
+  gitignore
+  gitmessage
+  hushlogin
+  khdrc
+  ripgreprc
+  rgignore
+  tern-project
+  tmux.conf
+  tmuxline_snapshot
+  vimrc
+  wgetrc
+  zshrc
+)
+
+for dotfile in ${dots[*]}; do
+  symlink "$dotfile" .
+done
+
+# Files that need slightly different symlink
+
+[ ! -e "$HOME/.tern-config" ] && ln -sfv $(grealpath tern-project) "$HOME/.tern-config"
+
 mkdir "$HOME/.ssh"
-ln -sfv $(grealpath ssh-config) "$HOME/.ssh/config"
+[ ! -e "$HOME/.ssh/config" ] && ln -sfv $(grealpath ssh-config) "$HOME/.ssh/config"
+
 popd
 
 # Directories
 
-# Remove first so folders don't become nested if this runs more than once
-rm "$HOME/.tmuxinator"
-rm "$HOME/applescript"
-rm "$HOME/bin"
-
 symlink applescript
 symlink bin
-symlink_dotfile tmuxinator
+symlink tmuxinator .
 
 mkdir "$HOME/.zdata"
 touch "$HOME/.zdata/.z"
